Ponerse al dia // Se ajusto el display name

+++ ProgramacionIV2025_VicenteGallardo/Assets/Scripts/Player.cs
@@ -203,7 +203,7 @@ public class Player : MonoBehaviour
                     movTurret.rotateSpd = item.value;
                     break;
                 case StatType.Attack:
-                    bullet.dmg = item.value;
+                    bullet.dmg = (int)item.value;
                     break;
                 case StatType.Defense:
                     break;
